Code refactoring to fit the PageCache abstraction
Summary: As part of the last diff, we introduced a new type of cache for RocksDB
called the page cache. It is at the level of pages from the disk.

This change, re-factor the block cache implementation to fit the new
abstraction.

- Make CacheTier implemend PageCache
- Drop Primary and Secondary tiers (it's just CacheTier now)
- Drop the wrapper code to fit the Cache acquire-release behavior
- Scrap VolatileCache and re-write it for PageCache interface
- Fix the tests to not use Block abstaction
- Re-factor the tests to be more readable
- Fix the block cache benchmark code
- Fix RocksDB benchmark integration
- Add a new type of prefix called page_cache_key_prefix and use it as the prefix
  for page cache

Test Plan: Run unit test, blockcache_test, microbench and RocksDB randomread
workload benchmark

Reviewers:

CC: leveldb@

Task ID: #8247246

Blame Rev: