Bug fix to get the unit tests to pass with RAMFS
Summary: Added a hack to ignore IOC when not available. This helps to pass unit
tests.

This fix is not official. This is to get the branch to working state. A more
portable fix will be applied to master and synced later.

Test Plan: Run unit tests

Reviewers:

CC: leveldb@

Task ID: #9092581

Blame Rev: