Add stat to measure pipeline latency
Summary: We adopt pipelined IO flow for writes. The design approach helps to
keep to latecy for insert by user (RocksDB in this case) to ~0 and helps scale
to 8+GB/s throughput which NVM can technically reach.

But the lock contention in the queue can increase the latency and will reflect
in RocksDB read code path. This stat measures the latency so we can optimize it
if necessary.

Test Plan: run unit tests

Reviewers:

Subscribers: leveldb

Tasks: 11096657

Blame Revision: